requirements-prod.txt
# Dependências específicas para produção
# Incluir requirements.txt base + dependências de produção

-r requirements.txt

# Servidor WSGI
gunicorn==23.0.0

# Banco de dados adicional
redis==5.2.1
django-redis==5.4.0

# Monitoramento e logs (opcional)
sentry-sdk[django]==2.19.2

# Segurança adicional
django-cors-headers==4.6.0

# Email em produção (opcional)
django-anymail==12.0

# Backup (opcional)
django-dbbackup==4.1.0
